+++ pkg/is-valid/README.md
@@ -1,0 +1,3 @@
+<b>IsValid</b>
+
+...
