+++ pkg/is-valid/databox_validator.go
@@ -1,0 +1,6 @@
+package is_valid
+
+// DataBoxValidator ...
+type DataBoxValidator interface {
+	IsValid() bool
+}
